[perf] Cache nixpkgs resolution (#2576)
## Summary

This caches resolutions for unlocked nixpkgs (the default) for 90 days
on the user's machine. Otherwise, everytime devbox tries to resolve
`github:NixOS/nixpkgs/nixpkgs-unstable` it will download a new version
of nixpkgs. This is slow (40+ seconds) and usually nothing changes.

### If a user wants to update nixpkgs there are two ways of doing it:

* `devbox update` without any arguments will update nixpkgs. This is
existing functionality.
* `devbox update nixpkgs [...pkgs]` will also update nixpkgs.

I'm not convinced updating nixpkgs when using `update` without arguments
is best, but I didn't want to break past functionality. It adds 40+
seconds to update operations when there is no package to update.

An alternative command syntax considered was `devbox update --nixpkgs`.
The benefit of doing this would be to avoid conflicts in the future.

I did not use `stdenv` because it already exists and it is a different
package.

### Bug fixes:

* Updating a specific package (e.g. `devbox update go`) no longer
updates nixpkgs

### Alternative approaches:

* We could try to search the local nix database to try to find existing
nixpkgs, but this seemed more trouble than it was worth.

## How was it tested?

* Created a new project and logged the cache being used.
* Replaced the cache file data with an old version, created a new
project and observed the old version used.
* Tried several permutations of `devbox update` with and without
packages

## Community Contribution License

All community contributions in this pull request are licensed to the
project
maintainers under the terms of the
[Apache 2 License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request, I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 License as
stated in
the
[Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).